Moved navigation bar out of dashboard because it is not exclusive to that page only
